+++ banyand/trace/block_writer.go
@@ -149,11 +149,8 @@ type blockWriter struct {
 	primaryBlockMetadata           primaryBlockMetadata
 	totalUncompressedSpanSizeBytes uint64
 	totalCount                     uint64
-	minTimestamp                   int64
 	totalMinTimestamp              int64
 	totalMaxTimestamp              int64
-	minTimestampLast               int64
-	maxTimestamp                   int64
 	totalBlocksCount               uint64
 	hasWrittenBlocks               bool
 }
@@ -169,9 +166,6 @@ func (bw *blockWriter) reset() {
 	} else {
 		bw.tagType.reset()
 	}
-	bw.minTimestampLast = 0
-	bw.minTimestamp = 0
-	bw.maxTimestamp = 0
 	bw.totalUncompressedSpanSizeBytes = 0
 	bw.totalCount = 0
 	bw.totalBlocksCount = 0
@@ -230,12 +224,10 @@ func (bw *blockWriter) mustWriteBlock(tid string, b *block) {
 	if tid < bw.tidLast {
 		logger.Panicf("the tid=%s cannot be smaller than the previously written tid=%s", tid, bw.tidLast)
 	}
-	hasWrittenBlocks := bw.hasWrittenBlocks
-	if !hasWrittenBlocks {
+	if !bw.hasWrittenBlocks {
 		bw.tidFirst = tid
 		bw.hasWrittenBlocks = true
 	}
-	isSeenTid := tid == bw.tidLast
 	bw.tidLast = tid
 
 	bm := generateBlockMetadata()
@@ -251,16 +243,6 @@ func (bw *blockWriter) mustWriteBlock(tid string, b *block) {
 	if bw.totalCount == 0 || tm.max > bw.totalMaxTimestamp {
 		bw.totalMaxTimestamp = tm.max
 	}
-	if !hasWrittenBlocks || tm.min < bw.minTimestamp {
-		bw.minTimestamp = tm.min
-	}
-	if !hasWrittenBlocks || tm.max > bw.maxTimestamp {
-		bw.maxTimestamp = tm.max
-	}
-	if isSeenTid && tm.min < bw.minTimestampLast {
-		logger.Panicf("the block for tid=%s cannot contain timestamp smaller than %d, but it contains timestamp %d", tid, bw.minTimestampLast, tm.min)
-	}
-	bw.minTimestampLast = tm.min
 
 	bw.totalUncompressedSpanSizeBytes += bm.uncompressedSpanSizeBytes
 	bw.totalCount += bm.count
@@ -280,8 +262,6 @@ func (bw *blockWriter) mustFlushPrimaryBlock(data []byte) {
 		bw.metaData = bw.primaryBlockMetadata.marshal(bw.metaData)
 	}
 	bw.hasWrittenBlocks = false
-	bw.minTimestamp = 0
-	bw.maxTimestamp = 0
 	bw.tidFirst = ""
 }
 
@@ -293,12 +273,10 @@ func (bw *blockWriter) mustWriteRawBlock(r *rawBlock) {
 	if bm.traceID < bw.tidLast {
 		logger.Panicf("the tid=%s cannot be smaller than the previously written tid=%s", bm.traceID, bw.tidLast)
 	}
-	hasWrittenBlocks := bw.hasWrittenBlocks
-	if !hasWrittenBlocks {
+	if !bw.hasWrittenBlocks {
 		bw.tidFirst = bm.traceID
 		bw.hasWrittenBlocks = true
 	}
-	isSeenTid := bm.traceID == bw.tidLast
 	bw.tidLast = bm.traceID
 	if bw.traceIDFilter != nil && bw.traceIDFilter.filter != nil {
 		bw.traceIDFilter.filter.Add(convert.StringToBytes(bm.traceID))
@@ -312,16 +290,6 @@ func (bw *blockWriter) mustWriteRawBlock(r *rawBlock) {
 	if bw.totalCount == 0 || tm.max > bw.totalMaxTimestamp {
 		bw.totalMaxTimestamp = tm.max
 	}
-	if !hasWrittenBlocks || tm.min < bw.minTimestamp {
-		bw.minTimestamp = tm.min
-	}
-	if !hasWrittenBlocks || tm.max > bw.maxTimestamp {
-		bw.maxTimestamp = tm.max
-	}
-	if isSeenTid && tm.min < bw.minTimestampLast {
-		logger.Panicf("the block for tid=%s cannot contain timestamp smaller than %d, but it contains timestamp %d", bm.traceID, bw.minTimestampLast, tm.min)
-	}
-	bw.minTimestampLast = tm.min
 
 	bw.totalUncompressedSpanSizeBytes += bm.uncompressedSpanSizeBytes
 	bw.totalCount += bm.count